+++ app/Http/Controllers/ProductController.php
@@ -56,6 +56,10 @@ class ProductController extends Controller
 
     public function delete(int $id)
     {
+        $product = $this->productService->getByCondition(['id' => $id]);
+        if(!$product) {
+            return response('Bu id ile kayıt bulunamadı...',400);
+        }
         $isDeleted = $this->productService->delete($id);
         if(!$isDeleted) {
             return response('Silme işlemi başarısız...',400);
